Add operational_items handling and plot/tuto fixes
Add explicit handling for 'operational_items' in weight correction factors and weight correction additions so operational item weights are applied to the weight_breakdown and operating_empty correctly. Clean up and refactor plot_pressure_coefficient_distribution (formatting, indexing/reshape expressions, and minor savefig/index fixes) to improve readability and consistency. Disable the aerodynamics surrogate in the tutorial by setting aerodynamics.settings.use_surrogate = False for deterministic VLM runs. Remove generated VnV HTML (BWB_Top_View.html) and rename the blended wing body aerodynamics test to VLM_aerodynamics_test.py; other test files updated (fuel_tank_volume_test.py, test_automatic_regression.py).
