+++ Tutorials/Analysis_Aerodynamic_Polars/aircraft_aerodynamics_analysis.py
@@ -887,7 +887,8 @@ def base_analysis(vehicle):
     geometry = RCAIDE.Framework.Analyses.Geometry.Geometry() 
     analyses.append(geometry)
   
-    aerodynamics   = RCAIDE.Framework.Analyses.Aerodynamics.Vortex_Lattice_Method()     
+    aerodynamics = RCAIDE.Framework.Analyses.Aerodynamics.Vortex_Lattice_Method()
+    aerodynamics.settings.use_surrogate = False
     analyses.append(aerodynamics)
     
     return analyses
